requirements-api.txt
fastapi>=0.110.0
uvicorn>=0.29.0
google-generativeai>=0.8.0
numpy>=1.26.0
pydantic>=2.6.0
scipy>=1.12.0
